# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tenet Audit — EU AI Act Audit Log API",
    "description": "Tamper-evident AI call logging, risk classification (Article 6 / Annex III), incident reporting (Article 73) and auditor-ready evidence packs (Article 12).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Tenant key (ta_live_…) or admin key (ta_admin_…)"
      }
    },
...
  },
  "paths": {
...
    "/v1/log-ai-call": {
      "post": {
        "tags": [
          "ai-calls"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "occurred_at",
                  "model_id",
                  "human_oversight"
                ],
                "properties": {
                  "occurred_at": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "model_id": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "model_version": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "prompt_version": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "system_purpose": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "data_sources": {
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "maxLength": 500,
                      "type": "string"
                    }
                  },
                  "user_consent": {
                    "type": "boolean"
                  },
                  "consent_ref": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "human_oversight": {
                    "description": "Article 14 — was a human in a position to oversee/intervene?",
                    "type": "boolean"
                  },
                  "oversight_actor": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "decision_output": {},
                  "input_summary": {
                    "description": "Prompt metadata — do NOT send raw personal data"
                  },
                  "risk_tier": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "prohibited"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "high"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "limited"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "minimal"
                        ]
                      }
                    ]
                  },
                  "end_user_ref": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "error": {},
                  "appeal": {},
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "seq",
                    "record_hash",
                    "prev_hash",
                    "recorded_at"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "seq": {
                      "type": "number"
                    },
                    "record_hash": {
                      "type": "string"
                    },
                    "prev_hash": {
                      "type": "string"
                    },
                    "recorded_at": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/log-ai-call/batch": {
      "post": {
        "tags": [
          "ai-calls"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "records"
                ],
                "properties": {
                  "records": {
                    "minItems": 1,
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "occurred_at",
                        "model_id",
                        "human_oversight"
                      ],
                      "properties": {
                        "occurred_at": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "model_id": {
                          "minLength": 1,
                          "maxLength": 200,
                          "type": "string"
                        },
                        "model_version": {
                          "maxLength": 200,
                          "type": "string"
                        },
                        "prompt_version": {
                          "maxLength": 200,
                          "type": "string"
                        },
                        "system_purpose": {
                          "maxLength": 500,
                          "type": "string"
                        },
                        "data_sources": {
                          "maxItems": 100,
                          "type": "array",
                          "items": {
                            "maxLength": 500,
                            "type": "string"
                          }
                        },
                        "user_consent": {
                          "type": "boolean"
                        },
                        "consent_ref": {
                          "maxLength": 500,
                          "type": "string"
                        },
                        "human_oversight": {
                          "description": "Article 14 — was a human in a position to oversee/intervene?",
                          "type": "boolean"
                        },
                        "oversight_actor": {
                          "maxLength": 200,
                          "type": "string"
                        },
                        "decision_output": {},
                        "input_summary": {
                          "description": "Prompt metadata — do NOT send raw personal data"
                        },
                        "risk_tier": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "prohibited"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "high"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "limited"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "minimal"
                              ]
                            }
                          ]
                        },
                        "end_user_ref": {
                          "maxLength": 200,
                          "type": "string"
                        },
                        "error": {},
                        "appeal": {},
                        "metadata": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "receipts"
                  ],
                  "properties": {
                    "receipts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "seq",
                          "record_hash",
                          "prev_hash",
                          "recorded_at"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "seq": {
                            "type": "number"
                          },
                          "record_hash": {
                            "type": "string"
                          },
                          "prev_hash": {
                            "type": "string"
                          },
                          "recorded_at": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/ai-calls": {
      "get": {
        "tags": [
          "ai-calls"
        ],
        "parameters": [
          {
            "schema": {
              "format": "date-time",
              "type": "string"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "format": "date-time",
              "type": "string"
            },
            "in": "query",
            "name": "to",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "model_id",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "system_purpose",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "prohibited"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "high"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "limited"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "minimal"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "risk_tier",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "human_oversight",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "has_error",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "end_user_ref",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "seq of the last item from previous page"
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "next_cursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "additionalProperties": true,
                        "type": "object",
                        "required": [
                          "id",
                          "seq",
                          "occurred_at",
                          "recorded_at",
                          "model_id",
                          "model_version",
                          "prompt_version",
                          "system_purpose",
                          "data_sources",
                          "user_consent",
                          "consent_ref",
                          "human_oversight",
                          "oversight_actor",
                          "decision_output",
                          "input_summary",
                          "risk_tier",
                          "end_user_ref",
                          "error",
                          "appeal",
                          "metadata",
                          "prev_hash",
                          "record_hash"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "seq": {
                            "type": "number"
                          },
                          "occurred_at": {
                            "type": "string"
                          },
                          "recorded_at": {
                            "type": "string"
                          },
                          "model_id": {
                            "type": "string"
                          },
                          "model_version": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "prompt_version": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "system_purpose": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "data_sources": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "user_consent": {
                            "anyOf": [
                              {
                                "type": "boolean"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "consent_ref": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "human_oversight": {
                            "type": "boolean"
                          },
                          "oversight_actor": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "decision_output": {},
                          "input_summary": {},
                          "risk_tier": {
                            "anyOf": [
                              {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "enum": [
                                      "prohibited"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "high"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "limited"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "minimal"
                                    ]
                                  }
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "end_user_ref": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "error": {},
                          "appeal": {},
                          "metadata": {},
                          "prev_hash": {
                            "type": "string"
                          },
                          "record_hash": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "next_cursor": {
                      "anyOf": [
                        {
                          "type": "number"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/ai-calls/{id}": {
      "get": {
        "tags": [
          "ai-calls"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "required": [
                    "id",
                    "seq",
                    "occurred_at",
                    "recorded_at",
                    "model_id",
                    "model_version",
                    "prompt_version",
                    "system_purpose",
                    "data_sources",
                    "user_consent",
                    "consent_ref",
                    "human_oversight",
                    "oversight_actor",
                    "decision_output",
                    "input_summary",
                    "risk_tier",
                    "end_user_ref",
                    "error",
                    "appeal",
                    "metadata",
                    "prev_hash",
                    "record_hash"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "seq": {
                      "type": "number"
                    },
                    "occurred_at": {
                      "type": "string"
                    },
                    "recorded_at": {
                      "type": "string"
                    },
                    "model_id": {
                      "type": "string"
                    },
                    "model_version": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "prompt_version": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "system_purpose": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data_sources": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "user_consent": {
                      "anyOf": [
                        {
                          "type": "boolean"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "consent_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "human_oversight": {
                      "type": "boolean"
                    },
                    "oversight_actor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "decision_output": {},
                    "input_summary": {},
                    "risk_tier": {
                      "anyOf": [
                        {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "prohibited"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "high"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "limited"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "minimal"
                              ]
                            }
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "end_user_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error": {},
                    "appeal": {},
                    "metadata": {},
                    "prev_hash": {
                      "type": "string"
                    },
                    "record_hash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ],
...
}